\name{bootstrap.wrapper.cont.inter.Y.and.X}
\alias{bootstrap.wrapper.cont.inter.Y.and.X}
\title{Internal function used for the parallel computation on the bootstrap sample}
\usage{
  bootstrap.wrapper.cont.inter.Y.and.X(x, data,
    model, var.inter, Outcome.model, Nboot)
}
\description{
  Internal function used for the parallel computation on
  the bootstrap sample
}
\keyword{internal}

