\docType{data}
\name{data.pathway}
\alias{data.pathway}
\title{Fictive data frame corresponding to the gene x pathway analysis for the example of case-control study.}
\format{A data frame of 0 and 1 indicating in which pathway (column)
each gene (rows) is included.}
\description{
  Fictive data frame corresponding to the gene x pathway
  analysis for the example of case-control study.
}
\keyword{datasets}

