\docType{data}
\name{data.pige}
\alias{data.pige}
\title{Sample data of a case-control study}
\format{A dataframe with 1000 rows and 134 columns}
\description{
  This fictive data set contains one environmental variable
  (var_int), 4 fixed covariates (cov) and SNP variables
  (coded 0,1,2).
}
\keyword{datasets}

