\docType{data}
\name{data.surv}
\alias{data.surv}
\title{Fictive data set for sample data of a survival analysis.}
\format{A dataframe with 66 rows and 887 columns}
\description{
  This fictive data set contains a time to event (TIME), an
  indicator variable (named EVENT) coded 0/1
  (censored/event), one environmental variable (var_int),
  and SNP variables.
}
\keyword{datasets}

