\docType{data}
\name{list.gene.snp.surv}
\alias{list.gene.snp.surv}
\title{Fictive list (for survival example) containing the names of the snp for each gene included in the studied pathways}
\format{A list containing the names of the SNPs belonging to each gene analysed.}
\description{
  Fictive list (for survival example) containing the names
  of the snp for each gene included in the studied pathways
}
\keyword{datasets}

