\name{permutation.wrapper.cat.inter}
\alias{permutation.wrapper.cat.inter}
\title{Internal function used for the parallel computation on the permutation sample}
\usage{
  permutation.wrapper.cat.inter(x, mat, data, model,
    var.inter, Outcome.model)
}
\description{
  Internal function used for the parallel computation on
  the permutation sample
}
\keyword{internal}

