\name{permutation.wrapper.cont.inter.Y.and.X}
\alias{permutation.wrapper.cont.inter.Y.and.X}
\title{Internal function used for the parallel computation on the permutation sample}
\usage{
  permutation.wrapper.cont.inter.Y.and.X(x, mat, data,
    model, var.inter, Outcome.model)
}
\description{
  Internal function used for the parallel computation on
  the permutation sample
}
\keyword{internal}

