% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.wrapper.cont.inter.Y.within.X.R
\name{permutation.wrapper.cont.inter.Y.within.X}
\alias{permutation.wrapper.cont.inter.Y.within.X}
\title{Internal function used for the parallel computation on the permutation sample}
\usage{
permutation.wrapper.cont.inter.Y.within.X(x, mat, data, model, var.inter,
  Outcome.model)
}
\description{
Internal function used for the parallel computation on the permutation sample
}
\keyword{internal}
