\name{permutation.wrapper.cont.Y.and.X}
\alias{permutation.wrapper.cont.Y.and.X}
\title{Internal function used for the parallel computation on the permutation sample}
\usage{
  permutation.wrapper.cont.Y.and.X(x, mat, data, model,
    Outcome.model)
}
\description{
  Internal function used for the parallel computation on
  the permutation sample
}
\keyword{internal}

