\docType{data}
\name{snp.gene.select}
\alias{snp.gene.select}
\title{Fictive data frame containing the names of the gene (first column) and the names of the SNP belonging to the SNPS (second column)}
\format{A data frame of 130 entries corresponding of the gene's name of each SNP.}
\description{
  Fictive data frame containing the names of the gene
  (first column) and the names of the SNP belonging to the
  SNPS (second column)
}
\keyword{datasets}

